Renamed and refactored as ShowerSimulator
